refactor: split core loop stall and wait cycle helpers
