+++ src/orchestrator/loop/core-loop/task-cycle.ts
@@ -1,14 +1,11 @@
 /**
- * Completion checks, stall handling, and task execution for a CoreLoop iteration.
+ * Completion checks and task execution for a CoreLoop iteration.
  */
 
-import { randomUUID } from "node:crypto";
 import * as path from "node:path";
 import type { Goal } from "../../../base/types/goal.js";
-import type { GapVector, GapHistoryEntry } from "../../../base/types/gap.js";
-import type { StallReport } from "../../../base/types/stall.js";
+import type { GapVector } from "../../../base/types/gap.js";
 import type { DriveScore } from "../../../base/types/drive.js";
-import type { WaitExpiryOutcome } from "../../../base/types/strategy.js";
 import { KnowledgeGraph } from "../../../platform/knowledge/knowledge-graph.js";
 import { loadDreamActivationState, mergeUniqueKnowledgeEntries } from "../../../platform/dream/dream-activation.js";
 import {
@@ -20,7 +17,6 @@ import {
   getMilestones,
   evaluatePace,
 } from "../../goal/milestone-evaluator.js";
-import { gatherStallEvidence } from "../stall-evidence.js";
 import { verifyWithTools } from "../verification-layer1.js";
 import { buildLoopToolContext } from "./preparation.js";
 import {
@@ -29,26 +25,13 @@ import {
 } from "../../execution/context/context-builder.js";
 import type { CapabilityAcquisitionOutcome } from "./capability.js";
 import type { CoreLoopEvidenceLedger } from "./evidence-ledger.js";
-import { ApprovalStore } from "../../../runtime/store/approval-store.js";
-import { buildWaitApprovalId } from "../../strategy/portfolio-rebalance.js";
-import type { WaitStrategyActivationContext } from "../../strategy/strategy-manager-base.js";
-import type { GapObservation } from "../../../base/types/time-horizon.js";
+export { detectStallsAndRebalance } from "./task-cycle-stall.js";
+export {
+  evaluateWaitStrategiesForObserveOnly,
+  type WaitStrategyObservationDecision,
+} from "./task-cycle-wait.js";
 
 // ─── Phase 5 ───
-
-const WAIT_APPROVAL_TIMEOUT_MS = 24 * 60 * 60 * 1000;
-const WAIT_APPROVAL_REMINDER_MS = 15 * 60 * 1000;
-
-type DimensionGapSample = {
-  normalized_gap: number;
-  timestamp: string;
-};
-
-function resolveGoalWorkspacePath(goal: Goal): string | undefined {
-  const constraint = goal.constraints.find((entry) => entry.startsWith("workspace_path:"));
-  const workspacePath = constraint?.slice("workspace_path:".length).trim();
-  return workspacePath || undefined;
-}
 
 /** Completion check + milestone deadline check.
  * Sets result.error on fatal failure, sets result.completionJudgment. */
@@ -131,463 +114,6 @@ export async function checkCompletionAndMilestones(
   }
 }
 
-// ─── Phase 6 ───
-
-/** Stall detection per-dimension and globally, plus portfolio rebalance. */
-export async function detectStallsAndRebalance(
-  ctx: PhaseCtx,
-  goalId: string,
-  goal: Goal,
-  result: LoopIterationResult,
-  stallActionHints?: StallActionHints
-): Promise<void> {
-  try {
-    const gapHistory = await ctx.deps.stateManager.loadGapHistory(goalId);
-    const gapHistoryByDimension = indexGapHistoryByDimension(goal, gapHistory);
-    const waitActivationContext = buildWaitStrategyActivationContext(
-      ctx,
-      goalId,
-      goal,
-      gapHistoryByDimension
-    );
-
-    // Gather tool-based workspace evidence for stall detection (Phase 6)
-    if (ctx.toolExecutor) {
-      try {
-        const workspacePath = resolveGoalWorkspacePath(goal);
-        const toolContext = {
-          cwd: workspacePath ?? process.cwd(),
-          goalId,
-          trustBalance: 0,
-          preApproved: true,
-          approvalFn: async () => false,
-        };
-        const evidence = await gatherStallEvidence(ctx.toolExecutor, toolContext, workspacePath);
-        result.toolStallEvidence = evidence;
-        if (!evidence.hasWorkspaceChanges) {
-          ctx.logger?.info("CoreLoop: stall evidence — no workspace changes detected", { goalId, toolErrors: evidence.toolErrors });
-        }
-      } catch {
-        // Non-fatal: evidence gathering failure does not block stall detection
-      }
-    }
-
-    // Suppression uses WaitStrategy.wait_until as the authoritative source of truth.
-    // task.plateau_until is a task-local mirror for consumers that do not load portfolios.
-    // Collect suppressed dimensions from all active WaitStrategies; skip those dims in stall loop.
-    const suppressedDimensions = new Set<string>();
-    if (ctx.deps.portfolioManager) {
-      try {
-        const portfolio = await ctx.deps.strategyManager.getPortfolio(goalId);
-        if (portfolio) {
-          for (const s of portfolio.strategies) {
-            if (s.state !== "active" || !ctx.deps.portfolioManager.isWaitStrategy(s)) continue;
-            const ws = s as Record<string, unknown>;
-            const waitUntil = typeof ws["wait_until"] === "string" ? ws["wait_until"] as string : null;
-            if (!ctx.deps.stallDetector.isSuppressed(waitUntil)) continue;
-            // Suppression is intentionally primary_dimension-only. Wait expiry,
-            // baseline capture, and outcome evaluation are all keyed to the
-            // canonical wait dimension rather than every target_dimension.
-            const primaryDim = typeof ws["primary_dimension"] === "string" ? ws["primary_dimension"] as string : null;
-            if (primaryDim) {
-              suppressedDimensions.add(primaryDim);
-              ctx.logger?.info("CoreLoop: stall detection suppressed for dimension by active WaitStrategy", {
-                goalId,
-                dimension: primaryDim,
-                waitUntil,
-              });
-              result.waitSuppressed = true;
-            }
-          }
-        }
-      } catch {
-        // Non-fatal: suppression check failure does not block stall detection
-      }
-    }
-
-    // Per-dimension stall check (skip dimensions suppressed by active WaitStrategies)
-    for (const dim of goal.dimensions) {
-      if (suppressedDimensions.has(dim.name)) continue;
-      const dimGapHistory = gapHistoryByDimension.get(dim.name) ?? [];
-
-      const stallReport = ctx.deps.stallDetector.checkDimensionStall(
-        goalId,
-        dim.name,
-        dimGapHistory
-      );
-
-      if (stallReport) {
-        result.stallDetected = true;
-        result.stallReport = stallReport;
-
-        // Predicted stalls are advisory — log but don't pivot/escalate
-        if (
-          stallReport.stall_type === "predicted_plateau" ||
-          stallReport.stall_type === "predicted_regression"
-        ) {
-          ctx.logger?.info(
-            `CoreLoop: early warning ${stallReport.stall_type} — monitoring, no pivot`,
-            { goalId },
-          );
-          continue;
-        }
-
-        const escalationLevel = await ctx.deps.stallDetector.getEscalationLevel(goalId, dim.name);
-        await applyStallAction(
-          ctx,
-          goalId,
-          goal,
-          dimGapHistory,
-          stallReport,
-          escalationLevel,
-          dim.name,
-          result,
-          "",
-          stallActionHints,
-          waitActivationContext
-        );
-        break;
-      }
-    }
-
-    // Global stall check
-    if (!result.stallDetected) {
-      await checkGlobalStall(
-        ctx,
-        goalId,
-        goal,
-        result,
-        gapHistoryByDimension,
-        suppressedDimensions,
-        stallActionHints,
-        waitActivationContext
-      );
-    }
-
-    // Portfolio: check rebalance after stall detection
-    if (ctx.deps.portfolioManager) {
-      await rebalancePortfolio(ctx, goalId, goal, waitActivationContext);
-    }
-  } catch (err) {
-    ctx.logger?.warn("CoreLoop: stall detection failed (non-fatal)", { error: err instanceof Error ? err.message : String(err) });
-  }
-}
-
-function indexGapHistoryByDimension(
-  goal: Goal,
-  gapHistory: GapHistoryEntry[]
-): Map<string, DimensionGapSample[]> {
-  const indexedHistory = new Map<string, DimensionGapSample[]>();
-
-  for (const dim of goal.dimensions) {
-    indexedHistory.set(dim.name, []);
-  }
-
-  for (const entry of gapHistory) {
-    const seenDimensions = new Set<string>();
-    for (const gap of entry.gap_vector) {
-      if (seenDimensions.has(gap.dimension_name)) continue;
-      seenDimensions.add(gap.dimension_name);
-
-      const dimHistory = indexedHistory.get(gap.dimension_name);
-      if (!dimHistory) {
-        continue;
-      }
-
-      const normalizedGap = {
-        normalized_gap: gap.normalized_weighted_gap ?? 1,
-        timestamp: entry.timestamp,
-      };
-      dimHistory.push(normalizedGap);
-    }
-  }
-
-  return indexedHistory;
-}
-
-function buildWaitStrategyActivationContext(
-  ctx: PhaseCtx,
-  goalId: string,
-  goal: Goal,
-  gapHistoryByDimension: ReadonlyMap<string, DimensionGapSample[]>
-): WaitStrategyActivationContext | undefined {
-  if (!ctx.timeHorizonEngine) {
-    return undefined;
-  }
-
-  const velocityByDimension = new Map<string, number>();
-  for (const [dimensionName, dimHistory] of gapHistoryByDimension.entries()) {
-    if (dimHistory.length < 2) continue;
-    const currentGap = dimHistory[dimHistory.length - 1]?.normalized_gap;
-    if (typeof currentGap !== "number" || !Number.isFinite(currentGap)) continue;
-    const observations: GapObservation[] = dimHistory.map((entry) => ({
-      timestamp: entry.timestamp,
-      normalizedGap: entry.normalized_gap,
-    }));
-    const pacing = ctx.timeHorizonEngine.evaluatePacing(
-      goalId,
-      currentGap,
-      goal.deadline ?? null,
-      observations
-    );
-    if (Number.isFinite(pacing.velocityPerHour)) {
-      velocityByDimension.set(dimensionName, pacing.velocityPerHour);
-    }
-  }
-
-  return {
-    getCurrentGap: (_goalId, dimension) => {
-      const history = gapHistoryByDimension.get(dimension);
-      return history && history.length > 0
-        ? history[history.length - 1]?.normalized_gap ?? null
-        : null;
-    },
-    canAffordWait: ({ strategy, waitHours, currentGap, initialGap, startedAt }) => {
-      const velocityPerHour = velocityByDimension.get(strategy.primary_dimension);
-      if (velocityPerHour === undefined) {
-        return false;
-      }
-      const budget = ctx.timeHorizonEngine!.getTimeBudget(
-        goal.deadline ?? null,
-        goal.created_at ?? startedAt,
-        currentGap,
-        initialGap,
-        velocityPerHour
-      );
-      return budget.canAffordWait(waitHours);
-    },
-  };
-}
-
-// ─── Shared stall-action helper ───
-
-/** Apply REFINE/PIVOT/ESCALATE logic for a detected stall (per-dimension or global).
- * @param dimHistory      Gap history slice used for analysis (single-dim or first-dim for global).
- * @param stallReport     The detected StallReport.
- * @param escalationLevel Current escalation level for the stall dimension.
- * @param incrementDimName Dimension name passed to incrementEscalation after handling.
- * @param logPrefix       Short prefix for log messages, e.g. "" or "global ".
- */
-async function applyStallAction(
-  ctx: PhaseCtx,
-  goalId: string,
-  goal: Goal,
-  dimHistory: DimensionGapSample[],
-  stallReport: StallReport,
-  escalationLevel: number,
-  incrementDimName: string,
-  result: LoopIterationResult,
-  logPrefix: string,
-  stallActionHints?: StallActionHints,
-  waitActivationContext?: WaitStrategyActivationContext
-): Promise<void> {
-  if (ctx.deps.learningPipeline) {
-    try {
-      await ctx.deps.learningPipeline.onStallDetected(goalId, stallReport);
-    } catch {
-      // non-fatal
-    }
-  }
-
-  const activeStrategyForRecord = await Promise.resolve(ctx.deps.strategyManager.getActiveStrategy(goalId)).catch(() => null);
-  const strategyIdForRecord = activeStrategyForRecord?.id ?? "unknown";
-
-  // M14-S2: analyze stall cause to determine REFINE/PIVOT/ESCALATE
-  // Falls back to PIVOT behavior when analyzeStallCause is unavailable
-  const analysis = ctx.deps.stallDetector.analyzeStallCause?.(dimHistory);
-  result.stallAnalysis = analysis;
-  const selectedAction = analysis?.recommended_action === "escalate"
-    ? "escalate"
-    : stallActionHints?.recommendedAction ?? analysis?.recommended_action ?? "pivot";
-
-  if (selectedAction === "continue") {
-    ctx.logger?.info(`CoreLoop: ${logPrefix}stall CONTINUE — replanning evidence prefers continuing current strategy`, {
-      goalId,
-      evidence: analysis?.evidence,
-    });
-  } else if (stallReport.suggested_cause === "information_deficit" && ctx.deps.goalRefiner) {
-    // Observation-failure stall: re-refine the leaf to get better dimensions
-    ctx.logger?.info(`CoreLoop: ${logPrefix}observation-failure stall — calling reRefineLeaf`, { goalId });
-    try {
-      await ctx.deps.goalRefiner.reRefineLeaf(goalId, stallReport.suggested_cause!);
-    } catch (reRefineErr) {
-      ctx.logger?.warn(`CoreLoop: ${logPrefix}reRefineLeaf failed (non-fatal)`, {
-        goalId,
-        err: reRefineErr instanceof Error ? reRefineErr.message : String(reRefineErr),
-      });
-    }
-  } else if (selectedAction === "refine") {
-    // REFINE: keep current strategy, just log and continue
-    ctx.logger?.info(`CoreLoop: ${logPrefix}stall REFINE — parameter_issue detected, keeping strategy`, {
-      goalId,
-      evidence: analysis?.evidence,
-    });
-  } else if (selectedAction === "escalate") {
-    // ESCALATE: set escalation level to max to trigger loop exit
-    ctx.logger?.warn(`CoreLoop: ${logPrefix}stall ESCALATE — goal_unreachable detected`, {
-      goalId,
-      evidence: analysis?.evidence,
-    });
-    await ctx.deps.strategyManager.onStallDetected(
-      goalId,
-      3,
-      goal.origin ?? "general",
-      waitActivationContext
-    );
-    result.pivotOccurred = true;
-  } else {
-    // PIVOT: switch strategy, but check pivot count limit first
-    const portfolio = await ctx.deps.strategyManager.getPortfolio(goalId);
-    const activeStrategy = portfolio?.strategies.find((s) => s.state === "active");
-    const pivotCount = activeStrategy?.pivot_count ?? 0;
-    const maxPivotCount = activeStrategy?.max_pivot_count ?? 2;
-
-    if (pivotCount >= maxPivotCount) {
-      // Auto-escalate when pivot limit reached
-      ctx.logger?.warn(`CoreLoop: ${logPrefix}stall auto-ESCALATE — pivot_count limit reached`, {
-        goalId,
-        pivotCount,
-        maxPivotCount,
-      });
-      await ctx.deps.strategyManager.onStallDetected(
-        goalId,
-        3,
-        goal.origin ?? "general",
-        waitActivationContext
-      );
-      result.pivotOccurred = true;
-    } else {
-      const newStrategy = await ctx.deps.strategyManager.onStallDetected(
-        goalId,
-        escalationLevel + 1,
-        goal.origin ?? "general",
-        waitActivationContext
-      );
-      if (newStrategy) {
-        result.pivotOccurred = true;
-        if (activeStrategy?.id) {
-          try {
-            await ctx.deps.strategyManager.incrementPivotCount(goalId, activeStrategy.id);
-          } catch {
-            // non-fatal
-          }
-        }
-      }
-    }
-  }
-
-  // M14-S3: Record decision (non-fatal)
-  if (ctx.deps.knowledgeManager) {
-    try {
-      const latestGap = dimHistory[dimHistory.length - 1]?.normalized_gap ?? 1;
-      const recordedDecision = selectedAction === "continue" ? "proceed" : selectedAction;
-      await ctx.deps.knowledgeManager.recordDecision({
-        id: randomUUID(),
-        goal_id: goalId,
-        goal_type: goal.origin ?? "general",
-        strategy_id: strategyIdForRecord,
-        hypothesis: activeStrategyForRecord?.hypothesis,
-        decision: recordedDecision,
-        context: {
-          gap_value: latestGap,
-          stall_count: stallReport.escalation_level,
-          cycle_count: dimHistory.length,
-          trust_score: 0,
-        },
-        outcome: "pending",
-        timestamp: new Date().toISOString(),
-        what_worked: [],
-        what_failed: [],
-        suggested_next: [],
-      });
-    } catch {
-      // non-fatal: never block the loop for decision recording
-    }
-  }
-
-  if (incrementDimName) {
-    await ctx.deps.stallDetector.incrementEscalation(goalId, incrementDimName);
-  }
-}
-
-/** Global stall detection: check all dimensions together, handle REFINE/PIVOT/ESCALATE. */
-async function checkGlobalStall(
-  ctx: PhaseCtx,
-  goalId: string,
-  goal: Goal,
-  result: LoopIterationResult,
-  gapHistoryByDimension: Map<string, DimensionGapSample[]>,
-  suppressedDimensions: ReadonlySet<string>,
-  stallActionHints?: StallActionHints,
-  waitActivationContext?: WaitStrategyActivationContext
-): Promise<void> {
-  const activeGapHistoryByDimension =
-    suppressedDimensions.size === 0
-      ? gapHistoryByDimension
-      : new Map(
-          Array.from(gapHistoryByDimension.entries()).filter(
-            ([dimensionName]) => !suppressedDimensions.has(dimensionName)
-          )
-        );
-  if (activeGapHistoryByDimension.size === 0) {
-    return;
-  }
-
-  const globalStall = ctx.deps.stallDetector.checkGlobalStall(goalId, activeGapHistoryByDimension);
-  if (!globalStall) return;
-
-  result.stallDetected = true;
-  result.stallReport = globalStall;
-
-  const firstActiveDimension =
-    goal.dimensions.find((dimension) => !suppressedDimensions.has(dimension.name))?.name
-    ?? activeGapHistoryByDimension.keys().next().value
-    ?? "";
-  const firstDimHistory = activeGapHistoryByDimension.get(firstActiveDimension) ?? [];
-  const firstDimName = firstActiveDimension;
-
-  // Pass escalationLevel=1 so that escalationLevel+1=2, preserving the original global PIVOT level
-  await applyStallAction(
-    ctx,
-    goalId,
-    goal,
-    firstDimHistory,
-    globalStall,
-    1,
-    firstDimName,
-    result,
-    "global ",
-    stallActionHints,
-    waitActivationContext
-  );
-}
-
-/** Portfolio rebalance: check for normal rebalance triggers after stall detection. */
-async function rebalancePortfolio(
-  ctx: PhaseCtx,
-  goalId: string,
-  goal: Goal,
-  waitActivationContext?: WaitStrategyActivationContext
-): Promise<void> {
-  if (!ctx.deps.portfolioManager) return;
-  try {
-    const rebalanceTrigger = await ctx.deps.portfolioManager.shouldRebalance(goalId);
-    if (rebalanceTrigger) {
-      const rebalanceResult = await ctx.deps.portfolioManager.rebalance(goalId, rebalanceTrigger);
-      if (rebalanceResult.new_generation_needed) {
-        await ctx.deps.strategyManager.onStallDetected(
-          goalId,
-          3,
-          goal.origin ?? "general",
-          waitActivationContext
-        );
-      }
-    }
-  } catch {
-    // Portfolio rebalance errors are non-fatal
-  }
-}
-
 // ─── Phase 6b ───
 
 /** Check dependency graph block.
@@ -627,226 +153,6 @@ export interface TaskGenerationHints {
 
 export interface StallActionHints {
   recommendedAction?: "continue" | "refine" | "pivot";
-}
-
-export interface WaitStrategyObservationDecision {
-  observeOnly: boolean;
-  newGenerationNeeded: boolean;
-  outcome: WaitExpiryOutcome | null;
-}
-
-interface PendingWaitOutcome {
-  strategyId: string;
-  outcome: WaitExpiryOutcome;
-}
-
-/**
- * Evaluate active WaitStrategies before task generation. A due or pending wait
- * is an observe-only iteration: the loop records status and returns without
- * handing work to AgentLoop.
- */
-export async function evaluateWaitStrategiesForObserveOnly(
-  ctx: PhaseCtx,
-  goalId: string,
-  goal: Goal,
-  result: LoopIterationResult
-): Promise<WaitStrategyObservationDecision> {
-  if (!ctx.deps.portfolioManager) {
-    return { observeOnly: false, newGenerationNeeded: false, outcome: null };
-  }
-
-  const waitActivationContext = buildWaitStrategyActivationContext(
-    ctx,
-    goalId,
-    goal,
-    indexGapHistoryByDimension(goal, await ctx.deps.stateManager.loadGapHistory(goalId))
-  );
-
-  try {
-    const portfolio = await ctx.deps.strategyManager.getPortfolio(goalId);
-    if (!portfolio) {
-      return { observeOnly: false, newGenerationNeeded: false, outcome: null };
-    }
-
-    let firstNotDue: PendingWaitOutcome | null = null;
-
-    for (const strategy of portfolio.strategies) {
-      if (strategy.state !== "active" || !ctx.deps.portfolioManager.isWaitStrategy(strategy)) {
-        continue;
-      }
-
-      const waitOutcome = await ctx.deps.portfolioManager.handleWaitStrategyExpiry(
-        goalId,
-        strategy.id,
-        waitActivationContext
-      );
-      if (!waitOutcome) {
-        continue;
-      }
-
-      if (waitOutcome.status === "not_due") {
-        firstNotDue ??= { strategyId: strategy.id, outcome: waitOutcome };
-        continue;
-      }
-
-      result.waitStrategyId = strategy.id;
-      result.waitExpiryOutcome = waitOutcome;
-      result.waitObserveOnly = true;
-      result.waitExpired = true;
-
-      if (waitOutcome.status === "approval_required") {
-        result.waitApprovalId = await persistWaitApprovalPending(ctx, goalId, strategy, waitOutcome);
-      }
-
-      const waitTrigger = waitOutcome.rebalance_trigger ?? null;
-      if (waitTrigger) {
-        const rebalanceResult = await ctx.deps.portfolioManager.rebalance(goalId, waitTrigger);
-        if (rebalanceResult.new_generation_needed) {
-          await ctx.deps.strategyManager.onStallDetected(
-            goalId,
-            3,
-            goal.origin ?? "general",
-            waitActivationContext
-          );
-          result.waitObserveOnly = false;
-          return { observeOnly: false, newGenerationNeeded: true, outcome: waitOutcome };
-        }
-      }
-      return { observeOnly: true, newGenerationNeeded: false, outcome: waitOutcome };
-    }
-
-    if (firstNotDue) {
-      result.waitStrategyId = firstNotDue.strategyId;
-      result.waitExpiryOutcome = firstNotDue.outcome;
-      result.waitObserveOnly = true;
-      result.waitSuppressed = true;
-      return { observeOnly: true, newGenerationNeeded: false, outcome: firstNotDue.outcome };
-    }
-  } catch (err) {
-    ctx.logger?.warn("CoreLoop: wait observation failed (non-fatal)", {
-      goalId,
-      error: err instanceof Error ? err.message : String(err),
-    });
-  }
-
-  return { observeOnly: false, newGenerationNeeded: false, outcome: null };
-}
-
-async function persistWaitApprovalPending(
-  ctx: PhaseCtx,
-  goalId: string,
-  strategy: { id: string; wait_until?: unknown },
-  outcome: WaitExpiryOutcome
-): Promise<string | undefined> {
-  const strategyId = strategy.id;
-  try {
-    const now = Date.now();
-    const approvalId = buildWaitApprovalId(goalId, strategyId);
-    const timeoutMs = WAIT_APPROVAL_TIMEOUT_MS;
-    const nextObserveAt = new Date(now + WAIT_APPROVAL_REMINDER_MS).toISOString();
-    const expiresAt = new Date(now + timeoutMs).toISOString();
-    const task = {
-      id: `wait:${strategyId}`,
-      description: outcome.details ?? "WaitStrategy requires approval before continuing",
-      action: "wait_strategy_resume_approval",
-    };
-
-    if (ctx.deps.waitApprovalBroker) {
-      void ctx.deps.waitApprovalBroker.requestApproval(goalId, task, timeoutMs, approvalId).catch((err) => {
-        ctx.logger?.warn("CoreLoop: wait approval broker request failed", {
-          goalId,
-          strategyId,
-          error: err instanceof Error ? err.message : String(err),
-        });
-      });
-      await postponeWaitObservationForApproval(ctx, goalId, strategy, approvalId, nextObserveAt, expiresAt);
-      return approvalId;
-    }
-
-    const baseDir = typeof ctx.deps.stateManager.getBaseDir === "function"
-      ? ctx.deps.stateManager.getBaseDir()
-      : null;
-    if (!baseDir) return undefined;
-
-    const approvalStore = new ApprovalStore(path.join(baseDir, "runtime"));
-    await approvalStore.savePending({
-      approval_id: approvalId,
-      goal_id: goalId,
-      request_envelope_id: approvalId,
-      correlation_id: approvalId,
-      state: "pending",
-      created_at: now,
-      expires_at: now + timeoutMs,
-      payload: {
-        task,
-        wait_strategy_id: strategyId,
-        wait_outcome: outcome,
-      },
-    });
-    await postponeWaitObservationForApproval(ctx, goalId, strategy, approvalId, nextObserveAt, expiresAt);
-    return approvalId;
-  } catch (err) {
-    ctx.logger?.warn("CoreLoop: failed to persist wait approval request", {
-      goalId,
-      strategyId,
-      error: err instanceof Error ? err.message : String(err),
-    });
-    return undefined;
-  }
-}
-
-async function postponeWaitObservationForApproval(
-  ctx: PhaseCtx,
-  goalId: string,
-  strategy: { id: string; wait_until?: unknown },
-  approvalId: string,
-  nextObserveAt: string,
-  expiresAt: string
-): Promise<void> {
-  try {
-    const path = `strategies/${goalId}/wait-meta/${strategy.id}.json`;
-    const raw = await ctx.deps.stateManager.readRaw(path);
-    const metadata = raw && typeof raw === "object" ? raw as Record<string, unknown> : {};
-    const waitUntil = typeof metadata["wait_until"] === "string"
-      ? metadata["wait_until"]
-      : typeof strategy.wait_until === "string"
-        ? strategy.wait_until
-        : nextObserveAt;
-    const conditions = Array.isArray(metadata["conditions"]) && metadata["conditions"].length > 0
-      ? metadata["conditions"]
-      : [{ type: "time_until", until: waitUntil }];
-
-    await ctx.deps.stateManager.writeRaw(path, {
-      ...metadata,
-      schema_version: 1,
-      wait_until: waitUntil,
-      conditions,
-      resume_plan: metadata["resume_plan"] ?? { action: "complete_wait" },
-      next_observe_at: nextObserveAt,
-      latest_observation: {
-        status: "pending",
-        evidence: {
-          approval_pending: true,
-          approval_id: approvalId,
-        },
-        next_observe_at: nextObserveAt,
-        confidence: 1,
-        resume_hint: "waiting_for_approval",
-      },
-      approval_pending: {
-        approval_id: approvalId,
-        requested_at: new Date().toISOString(),
-        next_reminder_at: nextObserveAt,
-        expires_at: expiresAt,
-      },
-    });
-  } catch (err) {
-    ctx.logger?.warn("CoreLoop: failed to postpone wait observation for approval", {
-      goalId,
-      strategyId: strategy.id,
-      error: err instanceof Error ? err.message : String(err),
-    });
-  }
 }
 
 /** Collect context, run task cycle, handle capability acquisition,
